Changes to OFDM_Tx_Rx_fixed_point model
Have removed fine timing synchronisation due to previous problems with thresholding. The timing synch is now based on the initial timing obtained using the Schmidl and Cox metric.

Changed baseband sampling rate from 200kHz to  1MHz. This is a more realistic sampling rate for OFDM systems. Modified the interpolation and decimation changes to accommodate this change.